requirements.txt
numpy==1.21.2
pandas==1.3.4
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
PyYAML==6.0
folium==0.12.0
joblib==1.1.0
notebook==6.4.6
protobuf==3.19.1
tensorflow==2.7.0
tensorflow-probability==0.15.0
gpflow==2.3.0
wheel==0.37.0
pipwin==0.5.2